requirements.txt
# Web app
streamlit==1.38.0

PyMuPDF==1.24.9
PyPDF2==3.0.1
pillow==10.4.0

# Google Vision OCR
google-cloud-vision==3.8.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utils
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
wordfreq==1.5